requirements.txt
django
djangorestframework
django_bootstrap4